+++ utils/dat2nc.c
@@ -147,7 +147,7 @@ int add_decomp(int ncid, const char *infname, int label) {
         else if (ndims == 3)
             printf("label D%d: dims = %lld x %lld x %lld (in C order)\n", label, dims[2], dims[1], dims[0]);
     }
-    dimX = dims[0]; /* the least significant dimension */
+    dimX = (int)dims[0]; /* the least significant dimension */
 
     /* gsize is total number of elements in the global array */
     gsize = dims[0];
